Unsupporting setCompositing for RPi3
:Release Notes:
Unsupporting setCompositing for RPi3

:Detailed Notes:
The drm driver does not support changing the zorder or alpha value of
the hw plane for rpi3.
So in case of val_rpi, modify it to return true when the function is
called.

:Testing Performed:
luna call test

:QA Notes:

:Issues Addressed:
[PLAT-79525] [RPI]:
setVideoData,display/setCompositing,display/setDisplayWindow method are
failing under com.webos.service.videooutput

Change-Id: Ie16417c16403e4be8d192714ea22c84cb373a8c1
Reviewed-on: http://gpro.lge.com/245769
Tested-by: Kwanghee Lee <ekwang.lee@lge.com>
Reviewed-by: Nakyeon Choi <nakyeon.choi@lge.com>
Reviewed-by: Bongkyu Jeon <jeon.bongkyu@lgepartner.com>
Reviewed-by: Jaehoon Lee <jaehoon85.lee@lge.com>
Reviewed-by: Kwanghee Lee <ekwang.lee@lge.com>